Replace wand by img2pdf
It turns out that wand.image.Image holds an uncompressed copy of the
image instead of the JPEG itself. This results in huge memomry
consumption (1Gb per 10 pages) and thus makes it impossible to scan
documents with a lot of pages.

We this change we gather the pages in memory a JPEG BLOBs and generate
the PDF with img2pdf which embedds the JPEG directly without decoding
them (or at least not all at the same time).

Signed-off-by: Konrad Vité <kvite@paktolos.net>